Move helper function to helpers.js
